requirements.txt
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
django-debug-toolbar